test(test): test.js에 테스트 코드 함수 분리후 정리
