+++ __tests__/AdditionalTest.js
@@ -1,97 +1,4 @@
-import { DateTimes, MissionUtils } from '@woowacourse/mission-utils';
-import { EOL as LINE_SEPARATOR } from 'os';
-import App from '../src/App.js';
-
-const mockQuestions = (inputs) => {
-  const messages = [];
-
-  MissionUtils.Console.readLineAsync = jest.fn((prompt) => {
-    messages.push(prompt);
-    const input = inputs.shift();
-
-    if (input === undefined) {
-      throw new Error('NO INPUT');
-    }
-
-    return Promise.resolve(input);
-  });
-
-  MissionUtils.Console.readLineAsync.messages = messages;
-};
-
-const mockNowDate = (date = null) => {
-  const mockDateTimes = jest.spyOn(MissionUtils.DateTimes, 'now');
-  mockDateTimes.mockReturnValue(new Date(date));
-  return mockDateTimes;
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
-  logSpy.mockClear();
-  return logSpy;
-};
-
-const getOutput = (logSpy) => [...logSpy.mock.calls].join(LINE_SEPARATOR);
-
-const expectLogContains = (received, expects) => {
-  expects.forEach((exp) => {
-    expect(received).toContain(exp);
-  });
-};
-
-const expectLogContainsWithoutSpacesAndEquals = (received, expects) => {
-  const processedReceived = received.replace(/[\s=]/g, '');
-  expects.forEach((exp) => {
-    expect(processedReceived).toContain(exp);
-  });
-};
-
-const runExceptions = async ({
-  inputs = [],
-  inputsToTerminate = [],
-  expectedErrorMessage = '',
-}) => {
-  // given
-  const logSpy = getLogSpy();
-  mockQuestions([...inputs, ...inputsToTerminate]);
-
-  // when
-  const app = new App();
-  await app.run();
-
-  // then
-  expect(logSpy).toHaveBeenCalledWith(
-    expect.stringContaining(expectedErrorMessage),
-  );
-};
-
-const run = async ({
-  inputs = [],
-  inputsToTerminate = [],
-  expected = [],
-  expectedIgnoringWhiteSpaces = [],
-}) => {
-  // given
-  const logSpy = getLogSpy();
-  mockQuestions([...inputs, ...inputsToTerminate]);
-
-  // when
-  const app = new App();
-  await app.run();
-
-  const output = getOutput(logSpy);
-
-  // then
-  if (expectedIgnoringWhiteSpaces.length > 0) {
-    expectLogContainsWithoutSpacesAndEquals(
-      output,
-      expectedIgnoringWhiteSpaces,
-    );
-  }
-  if (expected.length > 0) {
-    expectLogContains(output, expected);
-  }
-};
+import { run, runExceptions, mockNowDate } from '../src/Test/Test.js';
 
 const INPUTS_TO_TERMINATE = ['[비타민워터-1]', 'N', 'N'];
